+++ test/type/Complex.test.js
@@ -1,8 +1,8 @@
 // test data type Complex
 
 var assert = require('assert'),
-    Complex = require('../../lib/type/Complex'),
-    math = require('../../index')();
+    Unit = require('../../lib/type/Unit'),
+    Complex = require('../../lib/type/Complex');
 
 describe('Complex', function () {
 
@@ -195,9 +195,9 @@ describe('Complex', function () {
     it('should save polar coordinates input correctly', function() {
       var complex1 = Complex.fromPolar({r: 0, phi: 4});
       var complex2 = Complex.fromPolar({r: 5, phi: 0});
-      var complex3 = Complex.fromPolar({r: 1, phi: math.pi});
-      var complex4 = Complex.fromPolar({r: 3, phi: math.pi / 2});
-      var complex5 = Complex.fromPolar({r: 3, phi: -math.pi / 2});
+      var complex3 = Complex.fromPolar({r: 1, phi: Math.PI});
+      var complex4 = Complex.fromPolar({r: 3, phi: Math.PI / 2});
+      var complex5 = Complex.fromPolar({r: 3, phi: -Math.PI / 2});
       assertComplex(complex1, 0, 0);
       assertComplex(complex2, 5, 0);
       assert.equal(complex3.re, -1);
@@ -212,9 +212,9 @@ describe('Complex', function () {
     });
 
     it('should accept angle units for phi properly', function() {
-      var fromDeg = Complex.fromPolar(1, math.unit('90deg')),
-          fromRad = Complex.fromPolar(1, math.unit('0rad')),
-          fromGrad = Complex.fromPolar(1, math.unit('100grad'));
+      var fromDeg = Complex.fromPolar(1, new Unit(90, 'deg')),
+          fromRad = Complex.fromPolar(1, new Unit(0, 'rad')),
+          fromGrad = Complex.fromPolar(1, new Unit(100, 'grad'));
       assert.equal(fromDeg.im, 1);
       assert.equal(fromGrad.im, 1);
       assert.equal(fromRad.im, 0);