+++ test/function/matrix/diag.test.js
@@ -1,5 +1,6 @@
 var assert = require('assert'),
-    math = require('../../../index')();
+    math = require('../../../index')(),
+    bignumber = math.bignumber;
 
 describe('diag', function() {
 
@@ -31,6 +32,66 @@ describe('diag', function() {
   it('should throw an error in case of invalid k', function() {
     assert.throws(function () {math.diag([[1,2,3],[4,5,6]], 'a')}, /Second parameter in function diag must be an integer/);
     assert.throws(function () {math.diag([[1,2,3],[4,5,6]], 2.4)}, /Second parameter in function diag must be an integer/);
+  });
+
+  describe('bignumber', function () {
+    var array123 = [bignumber(1),bignumber(2),bignumber(3)];
+    var array123456 = [
+      [bignumber(1),bignumber(2),bignumber(3)],
+      [bignumber(4),bignumber(5),bignumber(6)]
+    ];
+
+    it('should return a diagonal matrix on the default diagonal', function() {
+      assert.deepEqual(math.diag(array123),
+          [
+            [bignumber(1),bignumber(0),bignumber(0)],
+            [bignumber(0),bignumber(2),bignumber(0)],
+            [bignumber(0),bignumber(0),bignumber(3)]
+          ]);
+
+      assert.deepEqual(math.diag(array123456), [bignumber(1),bignumber(5)]);
+    });
+
+    it('should return a array output on array input', function() {
+      assert.deepEqual(math.diag([bignumber(1),bignumber(2)]),
+          [
+            [bignumber(1),bignumber(0)],
+            [bignumber(0),bignumber(2)]
+          ]);
+    });
+
+    it('should return a matrix output on matrix input', function() {
+      assert.deepEqual(math.diag(math.matrix([bignumber(1),bignumber(2)])),
+          math.matrix([
+            [bignumber(1),bignumber(0)],
+            [bignumber(0),bignumber(2)]
+          ]));
+      assert.deepEqual(math.diag(math.matrix([
+        [bignumber(1),bignumber(2)],
+        [bignumber(3),bignumber(4)]
+      ])), math.matrix([bignumber(1),bignumber(4)]));
+    });
+
+    it('should put vector on given diagonal k in returned matrix', function() {
+      assert.deepEqual(math.diag(array123, bignumber(1)), [
+        [bignumber(0),bignumber(1),bignumber(0),bignumber(0)],
+        [bignumber(0),bignumber(0),bignumber(2),bignumber(0)],
+        [bignumber(0),bignumber(0),bignumber(0),bignumber(3)]
+      ]);
+      assert.deepEqual(math.diag(array123, bignumber(-1)), [
+        [bignumber(0),bignumber(0),bignumber(0)],
+        [bignumber(1),bignumber(0),bignumber(0)],
+        [bignumber(0),bignumber(2),bignumber(0)],
+        [bignumber(0),bignumber(0),bignumber(3)]
+      ]);
+    });
+
+    it('should return diagonal k from a matrix', function() {
+      assert.deepEqual(math.diag(array123456, bignumber(1)), [bignumber(2),bignumber(6)]);
+      assert.deepEqual(math.diag(array123456, bignumber(-1)), [bignumber(4)]);
+      assert.deepEqual(math.diag(array123456, bignumber(-2)), []);
+    });
+
   });
 
   it('should throw an error of the input matrix is not valid', function() {